requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & NLP
anthropic>=0.18.0
langchain>=0.1.0
python-dateutil>=2.8.2

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Cache & Queue
redis>=5.0.1
celery>=5.3.4

# HTTP Clients
httpx>=0.26.0
aiohttp>=3.9.1

# MCP Protocol
mcp>=0.1.0  # Model Context Protocol SDK

# Authentication & Security
PyJWT>=2.8.0
cryptography>=42.0.0
bcrypt>=4.1.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
python-dotenv>=1.0.0
black>=24.1.1
ruff>=0.1.14
mypy>=1.8.0

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
